+++ CHANGELOG.md
@@ -1,14 +1,28 @@
 GovReady-Q Release Notes
 ========================
 
-v.999 (December XX, 2020)
--------------------------
+Add default Organizational Defined Parameter values.
+
+**Developer changes**
+
+* Add default `controls.models.OrgParams` class to support basic, default generation of orgizational defined parameters.
+
+v.0.9.1.48.1 (December 17, 2020)
+--------------------------------
+
+**Bug fixes**
+
+* Fix handling of static files. Create new `static-root` directory outside of `siteapp` into which to collect static files.
+* Remove bad path reference to select2 javascript libraries in component library page.
+
+v.0.9.1.48 (December 15, 2020)
+------------------------------
 
 Add Component Library feature pages and improve UI for managing reuse and "certified" component library.
 
 Properly generate JSON, YAML questionnaire output documents from a JSON (or YAML) output template in the compliance app `output` section. The JSON, YAML output documents are first converted to Python data structures and then populated with information in a variant of Jinja2 substitutions.
 
-Add default Organizational Defined Parameter values.
+Fix tests so they execute successfully in CircleCI.
 
 **Feature changes**
 
@@ -24,6 +38,8 @@ Add default Organizational Defined Parameter values.
 * Support generation of JSON, YAML questionnaire output documents with Jinja2 style substitutions, loops, and conditionals. Re-do the 'json' template format to recognize a new %for control structure objects that execute loops.
 
 * Support generation of Word DOCX questionnaire output documents with page numbers, headers, footers, TOC (using pandoc custom reference doc feature).
+
+* Support creating a new component in the library.
 
 **UI changes**
 
@@ -54,11 +70,14 @@ Add default Organizational Defined Parameter values.
 
 * Fix multiple loadings of updated `smt.body` into bootstrap's panel heading section by improved naming of div classes in panel and better targeted update.
 * Fix enable_experimental_oscal control. Model method was set incorrectly requiring both enable_experimental_oscal and enable_experimental_opencontrol had to be enabled for either to show up.
+* Fix testing issues. Fix tests so they execute successfully in CircleCI.
 
 **Developer changes**
 
-* Add `custom-reference.docx` MS Word DOCX document to `/assets` directory to be used by pandoc when generating MS Word output documents in order to provide page numbers, headers, footers, TOC.
+* Default Selenium tests to headless mode. Add new `test_visible` parameter option for `local/environment.json` to force Selenium tests to run in visible or headless mode.
+Add `custom-reference.docx` MS Word DOCX document to `/assets` directory to be used by pandoc when generating MS Word output documents in order to provide page numbers, headers, footers, TOC.
 * Significantly refactored indentations in control edtor pages to make code folding and div analysis easier.
+* Add an ElementForm to create new components (AKA Elements).
 * Modified controls.Statement model to link `control_implementation` statements to
   `control_implementation_prototype` statements. See commit 5083af.
 * Add methods for diff'ing (e.g., comparing) a `control_implementation` statement against its prototype statement using Google diff-match-patch.
@@ -123,8 +142,7 @@ Example:
 ```
 * Update various libraries. See changes in `requirements.txt`.
 * Removed instance of using sys.stderr and replaced with logger for proper logging.
-* Add default `controls.models.OrgParams` class to support basic, default generation of orgizational defined parameters.
-
+* Fix tests so they execute successfully in CircleCI.
 
 **Other**